fix: change sort function
